Move the guts of Format to a helper function
